ref(normalization): Remove StoreNormalizeProcessor
