+++ relay-event-normalization/src/normalize/mod.rs
@@ -1,22 +1,9 @@
 use std::hash::Hash;
-use std::sync::Arc;
 
-use itertools::Itertools;
-use relay_base_schema::metrics::{MetricResourceIdentifier, MetricUnit};
-use relay_event_schema::processor::{
-    MaxChars, ProcessValue, ProcessingAction, ProcessingResult, ProcessingState, Processor,
-};
-use relay_event_schema::protocol::{
-    ClientSdkInfo, Event, EventId, EventType, Level, MetricSummaryMapping, NelContext,
-    ReplayContext, TraceContext, VALID_PLATFORMS,
-};
-use relay_protocol::{
-    Annotated, Empty, Error, FromValue, IntoValue, Meta, Object, Remark, RemarkType, RuleCondition,
-    Value,
-};
+use relay_base_schema::metrics::MetricUnit;
+use relay_event_schema::protocol::{Event, VALID_PLATFORMS};
+use relay_protocol::RuleCondition;
 use serde::{Deserialize, Serialize};
-
-use crate::StoreConfig;
 
 pub mod breakdowns;
 pub mod contexts;
@@ -91,93 +78,6 @@ pub fn is_valid_platform(platform: &str) -> bool {
     VALID_PLATFORMS.contains(&platform)
 }
 
-/// The processor that normalizes events for store.
-pub struct StoreNormalizeProcessor {
-    config: Arc<StoreConfig>,
-}
-
-impl StoreNormalizeProcessor {
-    /// Creates a new normalization processor.
-    pub fn new(config: Arc<StoreConfig>) -> Self {
-        StoreNormalizeProcessor { config }
-    }
-
-    /// Returns the SDK info from the config.
-    fn get_sdk_info(&self) -> Option<ClientSdkInfo> {
-        self.config.client.as_ref().and_then(|client| {
-            client
-                .splitn(2, '/')
-                .collect_tuple()
-                .or_else(|| client.splitn(2, ' ').collect_tuple())
-                .map(|(name, version)| ClientSdkInfo {
-                    name: Annotated::new(name.to_owned()),
-                    version: Annotated::new(version.to_owned()),
-                    ..Default::default()
-                })
-        })
-    }
-
-    fn normalize_metrics_summaries(&self, event: &mut Event) {
-        if event.ty.value() == Some(&EventType::Transaction) {
-            normalize_all_metrics_summaries(event);
-        }
-    }
-
-    fn normalize_trace_context(&self, event: &mut Event) {
-        if let Some(context) = event.context_mut::<TraceContext>() {
-            context.client_sample_rate = Annotated::from(self.config.client_sample_rate);
-        }
-    }
-
-    fn normalize_replay_context(&self, event: &mut Event) {
-        if let Some(ref mut contexts) = event.contexts.value_mut() {
-            if let Some(replay_id) = self.config.replay_id {
-                contexts.add(ReplayContext {
-                    replay_id: Annotated::new(EventId(replay_id)),
-                    other: Object::default(),
-                });
-            }
-        }
-    }
-
-    /// Infers the `EventType` from the event's interfaces.
-    fn infer_event_type(&self, event: &Event) -> EventType {
-        // The event type may be set explicitly when constructing the event items from specific
-        // items. This is DEPRECATED, and each distinct event type may get its own base class. For
-        // the time being, this is only implemented for transactions, so be specific:
-        if event.ty.value() == Some(&EventType::Transaction) {
-            return EventType::Transaction;
-        }
-        if event.ty.value() == Some(&EventType::UserReportV2) {
-            return EventType::UserReportV2;
-        }
-
-        // The SDKs do not describe event types, and we must infer them from available attributes.
-        let has_exceptions = event
-            .exceptions
-            .value()
-            .and_then(|exceptions| exceptions.values.value())
-            .filter(|values| !values.is_empty())
-            .is_some();
-
-        if has_exceptions {
-            EventType::Error
-        } else if event.csp.value().is_some() {
-            EventType::Csp
-        } else if event.hpkp.value().is_some() {
-            EventType::Hpkp
-        } else if event.expectct.value().is_some() {
-            EventType::ExpectCt
-        } else if event.expectstaple.value().is_some() {
-            EventType::ExpectStaple
-        } else if event.context::<NelContext>().is_some() {
-            EventType::Nel
-        } else {
-            EventType::Default
-        }
-    }
-}
-
 /// Replaces snake_case app start spans op with dot.case op.
 ///
 /// This is done for the affected React Native SDK versions (from 3 to 4.4).
@@ -206,45 +106,6 @@ pub fn normalize_app_start_spans(event: &mut Event) {
                     }
                 }
             }
-        }
-    }
-}
-
-/// Replaces all incoming metric identifiers in the metric summary with the correct MRI.
-///
-/// The reasoning behind this normalization, is that the SDK sends namespace-agnostic metric
-/// identifiers in the form `metric_type:metric_name@metric_unit` and those identifiers need to be
-/// converted to MRIs in the form `metric_type:metric_namespace/metric_name@metric_unit`.
-fn normalize_metrics_summary_mris(m: MetricSummaryMapping) -> MetricSummaryMapping {
-    m.into_iter()
-        .map(|(key, value)| match MetricResourceIdentifier::parse(&key) {
-            Ok(mri) => (mri.to_string(), value),
-            Err(err) => (
-                key,
-                Annotated::from_error(Error::invalid(err), value.0.map(IntoValue::into_value)),
-            ),
-        })
-        .collect()
-}
-
-/// Normalizes all the metrics summaries across the event payload.
-fn normalize_all_metrics_summaries(event: &mut Event) {
-    if let Some(metrics_summary) = event._metrics_summary.value_mut().as_mut() {
-        metrics_summary.update_value(normalize_metrics_summary_mris);
-    }
-
-    let Some(spans) = event.spans.value_mut() else {
-        return;
-    };
-
-    for span in spans.iter_mut() {
-        let metrics_summary = span
-            .value_mut()
-            .as_mut()
-            .and_then(|span| span._metrics_summary.value_mut().as_mut());
-
-        if let Some(ms) = metrics_summary {
-            ms.update_value(normalize_metrics_summary_mris);
         }
     }
 }
@@ -351,187 +212,25 @@ pub struct PerformanceScoreConfig {
     pub profiles: Vec<PerformanceScoreProfile>,
 }
 
-impl Processor for StoreNormalizeProcessor {
-    fn process_event(
-        &mut self,
-        event: &mut Event,
-        meta: &mut Meta,
-        state: &ProcessingState<'_>,
-    ) -> ProcessingResult {
-        event.process_child_values(self, state)?;
-
-        // Override internal attributes, even if they were set in the payload
-        let event_type = self.infer_event_type(event);
-        event.ty = Annotated::from(event_type);
-        event.project = Annotated::from(self.config.project_id);
-        event.key_id = Annotated::from(self.config.key_id.clone());
-        event.version = Annotated::from(self.config.protocol_version.clone());
-        event.grouping_config = self
-            .config
-            .grouping_config
-            .clone()
-            .map_or(Annotated::empty(), |x| {
-                FromValue::from_value(Annotated::<Value>::from(x))
-            });
-
-        // Validate basic attributes
-        relay_event_schema::processor::apply(&mut event.platform, |platform, _| {
-            if is_valid_platform(platform) {
-                Ok(())
-            } else {
-                Err(ProcessingAction::DeleteValueSoft)
-            }
-        })?;
-
-        // Default required attributes, even if they have errors
-        event.errors.get_or_insert_with(Vec::new);
-        event.id.get_or_insert_with(EventId::new);
-        event.platform.get_or_insert_with(|| "other".to_string());
-        event.logger.get_or_insert_with(String::new);
-        event.extra.get_or_insert_with(Object::new);
-        event.level.get_or_insert_with(|| match event_type {
-            EventType::Transaction => Level::Info,
-            _ => Level::Error,
-        });
-        if event.client_sdk.value().is_none() {
-            event.client_sdk.set_value(self.get_sdk_info());
-        }
-
-        if event.platform.as_str() == Some("java") {
-            if let Some(event_logger) = event.logger.value_mut().take() {
-                let shortened = shorten_logger(event_logger, meta);
-                event.logger.set_value(Some(shortened));
-            }
-        }
-
-        // Normalize connected attributes and interfaces
-        self.normalize_metrics_summaries(event);
-        self.normalize_trace_context(event);
-        self.normalize_replay_context(event);
-
-        Ok(())
-    }
-}
-
-/// If the logger is longer than [`MaxChars::Logger`], it returns a String with
-/// a shortened version of the logger. If not, the same logger is returned as a
-/// String. The resulting logger is always trimmed.
-///
-/// To shorten the logger, all extra chars that don't fit into the maximum limit
-/// are removed, from the beginning of the logger.  Then, if the remaining
-/// substring contains a `.` somewhere but in the end, all chars until `.`
-/// (exclusive) are removed.
-///
-/// Additionally, the new logger is prefixed with `*`, to indicate it was
-/// shortened.
-fn shorten_logger(logger: String, meta: &mut Meta) -> String {
-    let original_len = bytecount::num_chars(logger.as_bytes());
-    let trimmed = logger.trim();
-    let logger_len = bytecount::num_chars(trimmed.as_bytes());
-    if logger_len <= MaxChars::Logger.limit() {
-        if trimmed == logger {
-            return logger;
-        } else {
-            if trimmed.is_empty() {
-                meta.add_remark(Remark {
-                    ty: RemarkType::Removed,
-                    rule_id: "@logger:remove".to_owned(),
-                    range: Some((0, original_len)),
-                });
-            } else {
-                meta.add_remark(Remark {
-                    ty: RemarkType::Substituted,
-                    rule_id: "@logger:trim".to_owned(),
-                    range: None,
-                });
-            }
-            meta.set_original_length(Some(original_len));
-            return trimmed.to_string();
-        };
-    }
-
-    let mut tokens = trimmed.split("").collect_vec();
-    // Remove empty str tokens from the beginning and end.
-    tokens.pop();
-    tokens.reverse(); // Prioritize chars from the end of the string.
-    tokens.pop();
-
-    let word_cut = remove_logger_extra_chars(&mut tokens);
-    if word_cut {
-        remove_logger_word(&mut tokens);
-    }
-
-    tokens.reverse();
-    meta.add_remark(Remark {
-        ty: RemarkType::Substituted,
-        rule_id: "@logger:replace".to_owned(),
-        range: Some((0, logger_len - tokens.len())),
-    });
-    meta.set_original_length(Some(original_len));
-
-    format!("*{}", tokens.join(""))
-}
-
-/// Remove as many tokens as needed to match the maximum char limit defined in
-/// [`MaxChars::Logger`], and an extra token for the logger prefix. Returns
-/// whether a word has been cut.
-///
-/// A word is considered any non-empty substring that doesn't contain a `.`.
-fn remove_logger_extra_chars(tokens: &mut Vec<&str>) -> bool {
-    // Leave one slot of space for the prefix
-    let mut remove_chars = tokens.len() - MaxChars::Logger.limit() + 1;
-    let mut word_cut = false;
-    while remove_chars > 0 {
-        if let Some(c) = tokens.pop() {
-            if !word_cut && c != "." {
-                word_cut = true;
-            } else if word_cut && c == "." {
-                word_cut = false;
-            }
-        }
-        remove_chars -= 1;
-    }
-    word_cut
-}
-
-/// If the `.` token is present, removes all tokens from the end of the vector
-/// until `.`. If it isn't present, nothing is removed.
-fn remove_logger_word(tokens: &mut Vec<&str>) {
-    let mut delimiter_found = false;
-    for token in tokens.iter() {
-        if *token == "." {
-            delimiter_found = true;
-            break;
-        }
-    }
-    if !delimiter_found {
-        return;
-    }
-    while let Some(i) = tokens.last() {
-        if *i == "." {
-            break;
-        }
-        tokens.pop();
-    }
-}
 #[cfg(test)]
 mod tests {
     use chrono::{TimeZone, Utc};
     use insta::{assert_debug_snapshot, assert_json_snapshot};
+    use itertools::Itertools;
+    use relay_base_schema::events::EventType;
     use relay_base_schema::metrics::DurationUnit;
     use relay_base_schema::spans::SpanStatus;
-    use relay_event_schema::processor::process_value;
     use relay_event_schema::protocol::{
-        Context, ContextInner, Contexts, DebugImage, DebugMeta, Exception, Frame, Geo, IpAddr,
-        LenientString, LogEntry, MetricSummary, MetricsSummary, PairList, RawStacktrace, Request,
-        Span, SpanId, Stacktrace, TagEntry, Tags, TraceId, User, Values,
+        ClientSdkInfo, Context, ContextInner, Contexts, DebugImage, DebugMeta, EventId, Exception,
+        Frame, Geo, IpAddr, LenientString, Level, LogEntry, PairList, RawStacktrace, ReplayContext,
+        Request, Span, SpanId, Stacktrace, TagEntry, Tags, TraceContext, TraceId, User, Values,
     };
     use relay_protocol::{
-        assert_annotated_snapshot, get_path, get_value, ErrorKind, FromValue, SerializableAnnotated,
+        assert_annotated_snapshot, get_path, get_value, Annotated, Error, ErrorKind, FromValue,
+        Object, SerializableAnnotated, Value,
     };
     use serde_json::json;
     use similar_asserts::assert_eq;
-    use std::collections::BTreeMap;
     use uuid::Uuid;
 
     use crate::stacktrace::normalize_non_raw_frame;
@@ -541,12 +240,6 @@ mod tests {
     };
 
     use super::*;
-
-    impl Default for StoreNormalizeProcessor {
-        fn default() -> Self {
-            StoreNormalizeProcessor::new(Arc::new(StoreConfig::default()))
-        }
-    }
 
     #[test]
     fn test_merge_builtin_measurement_keys() {
@@ -647,11 +340,7 @@ mod tests {
             ..Event::default()
         });
 
-        let config = StoreConfig::default();
-        let mut processor = StoreNormalizeProcessor::new(Arc::new(config));
         normalize_event(&mut event, &NormalizationConfig::default());
-
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
 
         let ip_addr = get_value!(event.user.ip_address!);
         assert_eq!(ip_addr, &IpAddr("2.125.160.216".to_string()));
@@ -675,10 +364,7 @@ mod tests {
             ..Event::default()
         });
 
-        let config = StoreConfig::default();
-        let mut processor = StoreNormalizeProcessor::new(Arc::new(config));
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
 
         assert_eq!(Annotated::empty(), event.value().unwrap().user);
     }
@@ -691,12 +377,7 @@ mod tests {
         });
 
         let ip_address = IpAddr::parse("2.125.160.216").unwrap();
-        let config = StoreConfig {
-            client_ip: Some(ip_address.clone()),
-            ..StoreConfig::default()
-        };
 
-        let mut processor = StoreNormalizeProcessor::new(Arc::new(config));
         normalize_event(
             &mut event,
             &NormalizationConfig {
@@ -704,7 +385,6 @@ mod tests {
                 ..Default::default()
             },
         );
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
 
         let ip_addr = get_value!(event.user.ip_address!);
         assert_eq!(ip_addr, &IpAddr("2.125.160.216".to_string()));
@@ -761,16 +441,13 @@ mod tests {
 
     #[test]
     fn test_event_level_defaulted() {
-        let processor = &mut StoreNormalizeProcessor::default();
         let mut event = Annotated::new(Event::default());
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, processor, ProcessingState::root()).unwrap();
         assert_eq!(get_value!(event.level), Some(&Level::Error));
     }
 
     #[test]
     fn test_transaction_level_untouched() {
-        let processor = &mut StoreNormalizeProcessor::default();
         let mut event = Annotated::new(Event {
             ty: Annotated::new(EventType::Transaction),
             timestamp: Annotated::new(Utc.with_ymd_and_hms(1987, 6, 5, 4, 3, 2).unwrap().into()),
@@ -790,7 +467,6 @@ mod tests {
             ..Event::default()
         });
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, processor, ProcessingState::root()).unwrap();
         assert_eq!(get_value!(event.level), Some(&Level::Info));
     }
 
@@ -804,9 +480,7 @@ mod tests {
             ..Event::default()
         });
 
-        let mut processor = StoreNormalizeProcessor::default();
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
 
         let event = event.value().unwrap();
 
@@ -825,9 +499,7 @@ mod tests {
             ..Event::default()
         });
 
-        let mut processor = StoreNormalizeProcessor::default();
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
 
         let event = event.value().unwrap();
 
@@ -842,9 +514,7 @@ mod tests {
             ..Event::default()
         });
 
-        let mut processor = StoreNormalizeProcessor::default();
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
         assert_eq!(get_value!(event.environment), None);
     }
     #[test]
@@ -854,13 +524,13 @@ mod tests {
             contexts: Annotated::new(Contexts(Object::new())),
             ..Event::default()
         });
-        let config = StoreConfig {
-            replay_id: Some(replay_id),
-            ..StoreConfig::default()
-        };
-        let mut processor = StoreNormalizeProcessor::new(Arc::new(config));
-        normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
+        normalize_event(
+            &mut event,
+            &NormalizationConfig {
+                replay_id: Some(replay_id),
+                ..Default::default()
+            },
+        );
 
         let event = event.value().unwrap();
 
@@ -881,9 +551,7 @@ mod tests {
             ..Event::default()
         });
 
-        let mut processor = StoreNormalizeProcessor::default();
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
 
         let environment = get_path!(event.environment!);
         let expected_original = &Value::String("none".to_string());
@@ -906,9 +574,7 @@ mod tests {
             ..Event::default()
         });
 
-        let mut processor = StoreNormalizeProcessor::default();
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
 
         let release = get_path!(event.release!);
         let expected_original = &Value::String("Latest".to_string());
@@ -939,9 +605,7 @@ mod tests {
             ..Event::default()
         });
 
-        let mut processor = StoreNormalizeProcessor::default();
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
 
         assert_eq!(get_value!(event.site), None);
         assert_eq!(get_value!(event.server_name), None);
@@ -993,9 +657,7 @@ mod tests {
             ..Event::default()
         });
 
-        let mut processor = StoreNormalizeProcessor::default();
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
 
         assert_eq!(get_value!(event.tags!).len(), 1);
     }
@@ -1020,9 +682,7 @@ mod tests {
             ..Event::default()
         });
 
-        let mut processor = StoreNormalizeProcessor::default();
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
 
         assert_eq!(
             get_value!(event.tags!),
@@ -1071,9 +731,7 @@ mod tests {
             ..Event::default()
         });
 
-        let mut processor = StoreNormalizeProcessor::default();
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
 
         // should keep the first occurrence of every tag
         assert_eq!(
@@ -1108,12 +766,7 @@ mod tests {
             contexts: Annotated::new(Contexts(object)),
             ..Event::default()
         });
-        let config = StoreConfig {
-            ..StoreConfig::default()
-        };
-        let mut processor = StoreNormalizeProcessor::new(Arc::new(config));
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
 
         let event = event.value().unwrap();
 
@@ -1134,9 +787,7 @@ mod tests {
             ..Event::default()
         });
 
-        let mut processor = StoreNormalizeProcessor::default();
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
 
         assert_eq!(
             get_path!(event.debug_meta!),
@@ -1179,8 +830,7 @@ mod tests {
             ..Frame::default()
         });
 
-        let mut processor = StoreNormalizeProcessor::default();
-        process_value(&mut frame, &mut processor, ProcessingState::root()).unwrap();
+        normalize_non_raw_frame(&mut frame);
 
         let frame = frame.value().unwrap();
         assert_eq!(frame.context_line.as_str(), Some("some line"));
@@ -1230,9 +880,7 @@ mod tests {
         ..Event::default()
     });
 
-        let mut processor = StoreNormalizeProcessor::default();
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
 
         assert_eq!(
             get_value!(event.tags!),
@@ -1262,9 +910,7 @@ mod tests {
 
         let mut event = Annotated::<Event>::from_json(json).unwrap();
 
-        let mut processor = StoreNormalizeProcessor::default();
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
 
         let dist = &event.value().unwrap().dist;
         let result = &Annotated::<String>::from_error(
@@ -1297,9 +943,7 @@ mod tests {
             ..Event::default()
         });
 
-        let mut processor = StoreNormalizeProcessor::default();
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
 
         assert_eq!(
             get_value!(event.exceptions.values[0].stacktrace!),
@@ -1319,13 +963,14 @@ mod tests {
     #[test]
     fn test_parses_sdk_info_from_header() {
         let mut event = Annotated::new(Event::default());
-        let mut processor = StoreNormalizeProcessor::new(Arc::new(StoreConfig {
-            client: Some("_fooBar/0.0.0".to_string()),
-            ..StoreConfig::default()
-        }));
 
-        normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
+        normalize_event(
+            &mut event,
+            &NormalizationConfig {
+                client: Some("_fooBar/0.0.0".to_string()),
+                ..Default::default()
+            },
+        );
 
         assert_eq!(
             get_path!(event.client_sdk!),
@@ -1344,11 +989,8 @@ mod tests {
             ..Default::default()
         });
 
-        let mut processor = StoreNormalizeProcessor::default();
-
         validate_event_timestamps(&mut event, &EventValidationConfig::default()).unwrap();
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
 
         assert_eq!(get_value!(event.received!), get_value!(event.timestamp!));
     }
@@ -1363,16 +1005,16 @@ mod tests {
             ..Default::default()
         });
 
-        let mut processor = StoreNormalizeProcessor::new(Arc::new(StoreConfig {
-            grouping_config: Some(json!({
-                "id": "legacy:1234-12-12".to_string(),
-            })),
-            ..Default::default()
-        }));
-
         validate_event_timestamps(&mut event, &EventValidationConfig::default()).unwrap();
-        normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
+        normalize_event(
+            &mut event,
+            &NormalizationConfig {
+                grouping_config: Some(json!({
+                    "id": "legacy:1234-12-12".to_string(),
+                })),
+                ..Default::default()
+            },
+        );
 
         insta::assert_ron_snapshot!(SerializableAnnotated(&event), {
             ".event_id" => "[event-id]",
@@ -1416,9 +1058,7 @@ mod tests {
 "#;
         let mut event = Annotated::<Event>::from_json(json).unwrap();
 
-        let mut processor = StoreNormalizeProcessor::default();
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
 
         assert_json_snapshot!(SerializableAnnotated(&event), {".received" => "[received]"}, @r#"
         {
@@ -1465,12 +1105,6 @@ mod tests {
         let max_secs_in_past = Some(30 * 24 * 3600);
         let max_secs_in_future = Some(60);
 
-        let mut processor = StoreNormalizeProcessor::new(Arc::new(StoreConfig {
-            received_at,
-            max_secs_in_past,
-            max_secs_in_future,
-            ..Default::default()
-        }));
         validate_transaction(&event, &TransactionValidationConfig::default()).unwrap();
         validate_event_timestamps(
             &mut event,
@@ -1482,7 +1116,6 @@ mod tests {
         )
         .unwrap();
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
 
         insta::assert_ron_snapshot!(SerializableAnnotated(&event), {
         ".event_id" => "[event-id]",
@@ -1525,12 +1158,6 @@ mod tests {
         let max_secs_in_past = Some(30 * 24 * 3600);
         let max_secs_in_future = Some(60);
 
-        let mut processor = StoreNormalizeProcessor::new(Arc::new(StoreConfig {
-            received_at,
-            max_secs_in_past,
-            max_secs_in_future,
-            ..Default::default()
-        }));
         validate_event_timestamps(
             &mut event,
             &EventValidationConfig {
@@ -1541,7 +1168,6 @@ mod tests {
         )
         .unwrap();
         normalize_event(&mut event, &NormalizationConfig::default());
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
 
         insta::assert_ron_snapshot!(SerializableAnnotated(&event), {
         ".event_id" => "[event-id]",
@@ -1584,8 +1210,7 @@ mod tests {
             .into(),
         );
 
-        let mut processor = StoreNormalizeProcessor::default();
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
+        normalize_event(&mut event, &NormalizationConfig::default());
         assert_annotated_snapshot!(event);
     }
 
@@ -1600,8 +1225,7 @@ mod tests {
             .into(),
         );
 
-        let mut processor = StoreNormalizeProcessor::default();
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
+        normalize_event(&mut event, &NormalizationConfig::default());
         assert_annotated_snapshot!(event);
     }
 
@@ -1616,8 +1240,7 @@ mod tests {
             .into(),
         );
 
-        let mut processor = StoreNormalizeProcessor::default();
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
+        normalize_event(&mut event, &NormalizationConfig::default());
         assert_annotated_snapshot!(event);
     }
 
@@ -1632,8 +1255,7 @@ mod tests {
             .into(),
         );
 
-        let mut processor = StoreNormalizeProcessor::default();
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
+        normalize_event(&mut event, &NormalizationConfig::default());
         assert_annotated_snapshot!(event);
     }
 
@@ -1648,8 +1270,7 @@ mod tests {
             .into(),
         );
 
-        let mut processor = StoreNormalizeProcessor::default();
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
+        normalize_event(&mut event, &NormalizationConfig::default());
         assert_annotated_snapshot!(event);
     }
 
@@ -1664,8 +1285,7 @@ mod tests {
             .into(),
         );
 
-        let mut processor = StoreNormalizeProcessor::default();
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
+        normalize_event(&mut event, &NormalizationConfig::default());
         assert_annotated_snapshot!(event);
     }
 
@@ -1682,8 +1302,7 @@ mod tests {
             .into(),
         );
 
-        let mut processor = StoreNormalizeProcessor::default();
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
+        normalize_event(&mut event, &NormalizationConfig::default());
         assert_annotated_snapshot!(event);
     }
 
@@ -1698,8 +1317,7 @@ mod tests {
             .into(),
         );
 
-        let mut processor = StoreNormalizeProcessor::default();
-        process_value(&mut event, &mut processor, ProcessingState::root()).unwrap();
+        normalize_event(&mut event, &NormalizationConfig::default());
         assert_annotated_snapshot!(event);
     }
 
@@ -2044,173 +1662,6 @@ mod tests {
                 other: {},
             },
         ]
-        "###);
-    }
-
-    #[test]
-    fn test_normalize_metrics_summary_metric_identifiers() {
-        let mut metric_tags = BTreeMap::new();
-        metric_tags.insert(
-            "transaction".to_string(),
-            Annotated::new("/hello".to_string()),
-        );
-
-        let metric_summary = vec![Annotated::new(MetricSummary {
-            min: Annotated::new(1.0),
-            max: Annotated::new(20.0),
-            sum: Annotated::new(21.0),
-            count: Annotated::new(2),
-            tags: Annotated::new(metric_tags),
-        })];
-
-        let mut metrics_summary = BTreeMap::new();
-        metrics_summary.insert(
-            "d:page_duration@millisecond".to_string(),
-            Annotated::new(metric_summary.clone()),
-        );
-        metrics_summary.insert(
-            "c:custom/page_click@none".to_string(),
-            Annotated::new(Default::default()),
-        );
-        metrics_summary.insert(
-            "s:user@none".to_string(),
-            Annotated::new(metric_summary.clone()),
-        );
-        metrics_summary.insert("invalid".to_string(), Annotated::new(metric_summary));
-        metrics_summary.insert(
-            "g:page_load@second".to_string(),
-            Annotated::new(Default::default()),
-        );
-        let metrics_summary = MetricsSummary(metrics_summary);
-
-        let mut event = Annotated::new(Event {
-            spans: Annotated::new(vec![Annotated::new(Span {
-                op: Annotated::new("my_span".to_owned()),
-                _metrics_summary: Annotated::new(metrics_summary.clone()),
-                ..Default::default()
-            })]),
-            _metrics_summary: Annotated::new(metrics_summary),
-            ..Default::default()
-        });
-        normalize_all_metrics_summaries(event.value_mut().as_mut().unwrap());
-        assert_json_snapshot!(SerializableAnnotated(&event), @r###"
-        {
-          "spans": [
-            {
-              "op": "my_span",
-              "_metrics_summary": {
-                "c:custom/page_click@none": [],
-                "d:custom/page_duration@millisecond": [
-                  {
-                    "min": 1.0,
-                    "max": 20.0,
-                    "sum": 21.0,
-                    "count": 2,
-                    "tags": {
-                      "transaction": "/hello"
-                    }
-                  }
-                ],
-                "g:custom/page_load@second": [],
-                "invalid": null,
-                "s:custom/user@none": [
-                  {
-                    "min": 1.0,
-                    "max": 20.0,
-                    "sum": 21.0,
-                    "count": 2,
-                    "tags": {
-                      "transaction": "/hello"
-                    }
-                  }
-                ]
-              }
-            }
-          ],
-          "_metrics_summary": {
-            "c:custom/page_click@none": [],
-            "d:custom/page_duration@millisecond": [
-              {
-                "min": 1.0,
-                "max": 20.0,
-                "sum": 21.0,
-                "count": 2,
-                "tags": {
-                  "transaction": "/hello"
-                }
-              }
-            ],
-            "g:custom/page_load@second": [],
-            "invalid": null,
-            "s:custom/user@none": [
-              {
-                "min": 1.0,
-                "max": 20.0,
-                "sum": 21.0,
-                "count": 2,
-                "tags": {
-                  "transaction": "/hello"
-                }
-              }
-            ]
-          },
-          "_meta": {
-            "_metrics_summary": {
-              "invalid": {
-                "": {
-                  "err": [
-                    [
-                      "invalid_data",
-                      {
-                        "reason": "failed to parse metric"
-                      }
-                    ]
-                  ],
-                  "val": [
-                    {
-                      "count": 2,
-                      "max": 20.0,
-                      "min": 1.0,
-                      "sum": 21.0,
-                      "tags": {
-                        "transaction": "/hello"
-                      }
-                    }
-                  ]
-                }
-              }
-            },
-            "spans": {
-              "0": {
-                "_metrics_summary": {
-                  "invalid": {
-                    "": {
-                      "err": [
-                        [
-                          "invalid_data",
-                          {
-                            "reason": "failed to parse metric"
-                          }
-                        ]
-                      ],
-                      "val": [
-                        {
-                          "count": 2,
-                          "max": 20.0,
-                          "min": 1.0,
-                          "sum": 21.0,
-                          "tags": {
-                            "transaction": "/hello"
-                          }
-                        }
-                      ]
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
         "###);
     }
 }
