+++ tests/integration/fixtures/security_report/csp_chrome.no_processing.output.json
@@ -23,12 +23,18 @@
   },
   "culprit": "style-src cdn.example.com",
   "environment": "production",
+  "grouping_config": {
+    "enhancements": "eJybzDhxY05qemJypZWRgaGlroGxrqHRBABbEwcC",
+    "id": "legacy:2019-03-12"
+  },
+  "key_id": "123",
   "level": "error",
   "logentry": {
     "formatted": "Blocked 'style' from 'localhost:8000'"
   },
   "logger": "csp",
   "platform": "other",
+  "project": 42,
   "release": "01d5c3165d9fbc5c8bdcf9550a1d6793a80fc02b",
   "request": {
     "headers": [
@@ -51,5 +57,6 @@
   ],
   "user": {
     "ip_address": "127.0.0.1"
-  }
+  },
+  "version": "7"
 }
