+++ tests/integration/fixtures/security_report/hpkp.no_processing.output.json
@@ -18,8 +18,14 @@
   },
   "logger": "csp",
   "platform": "other",
+  "project": 42,
   "release": "01d5c3165d9fbc5c8bdcf9550a1d6793a80fc02b",
   "environment": "production",
+  "grouping_config": {
+    "enhancements": "eJybzDhxY05qemJypZWRgaGlroGxrqHRBABbEwcC",
+    "id": "legacy:2019-03-12"
+  },
+  "key_id": "123",
   "request": {
     "headers": [
       [
@@ -62,5 +68,6 @@
   },
   "user": {
     "ip_address": "127.0.0.1"
-  }
+  },
+  "version": "7"
 }
